narrows coordinate types from u64/i64 to u32
FiberAnnotation.pos and ref_pos change from i64/Option<i64> to
u32/Option<u32>.  seq_len and align_len on CurrRead narrow from
i64/u64 to u32.  GenomicRegion, WindowState, and several region/
interval helpers now store coordinates as u32.  Genomic positions
and read lengths fit comfortably in u32 and this reduces per-
annotation memory use.

Because u32 coordinates always fit in i64, the conversion from
GenomicRegion to bam::FetchDefinition becomes a plain From impl
rather than a fallible TryFrom.  Contigs longer than u32::MAX are
rejected by region-to-BED conversion checks, and simulate_mod_bam.rs
applies the same guard where needed.

ModTableEntry keeps ref_pos as i64 to accommodate the -1 sentinel
for unmapped positions while still covering the full u32 range.
Error::InvalidRegion fields pos and contig_length narrow to u32.
AlignmentInfo is added to the public re-exports in src/lib.rs.

filter_mods_by_ref_pos now accepts u32 start/end.  The Ordering::Equal
branch in set_mod_data_restricted_options passes start==end through
directly, stripping all mod data, replacing the former i64::MAX-1
sentinel pair.  ranges_filter_mods_by_ref_pos_with_start_equals_end
is updated to assert an empty result rather than expect an error.

seq_coords_from_ref_coords_overflowing_region_errors is renamed to
seq_coords_from_ref_coords_overflowing_read_errors: the test now
constructs a read with start at u32::MAX and checks that a region
intersection produces an error (the former i64::MAX overflow is no
longer reachable).

Adds test point_annotation_is_included_at_its_own_position, a small
reads_table.rs refactor around comment stripping, and a guard for a
for loop in reads_table.rs.

CLAUDE.md now instructed to look for specific versions of clang
if needed

+++ src/cli.rs
@@ -44,8 +44,8 @@ use std::str::FromStr;
 ///
 /// let bam = InputBamBuilder::default()
 ///     .bam_path(PathOrURLOrStdin::Path("/some/path/to/bam.bam".into()))
-///     .min_seq_len(30000u64)
-///     .min_align_len(20000i64)
+///     .min_seq_len(30000u32)
+///     .min_align_len(20000u32)
 ///     .read_id("some-id")
 ///     .read_filter("primary_forward,secondary_forward".into())
 ///     .sample_fraction(F32Bw0and1::new(1.0).expect("no error"))
@@ -115,7 +115,7 @@ use std::str::FromStr;
 ///     .bam_path(PathOrURLOrStdin::Path("/some/path/to/bam.bam".into()))
 ///     .read_id("some-id")
 ///     .region("chr4:1000-2000".into())
-///     .region_bed3(Bed3::<i32,u64>::new(3, 1000, 2000))
+///     .region_bed3(Bed3::<i32,u32>::new(3, 1000, 2000))
 ///     .build()?;
 /// # Ok::<(), Error>(())
 /// ```
@@ -165,12 +165,12 @@ pub struct InputBam {
     /// below this value. Defaults to 0.
     #[clap(long, default_value_t)]
     #[builder(setter(into))]
-    pub min_seq_len: u64,
+    pub min_seq_len: u32,
     /// Exclude reads whose alignment length in the BAM file is
     /// below this value. Defaults to unused.
     #[clap(long)]
     #[builder(setter(into, strip_option))]
-    pub min_align_len: Option<i64>,
+    pub min_align_len: Option<u32>,
     /// Only include this read id, defaults to unused i.e. all reads are used.
     /// NOTE: if there are multiple alignments corresponding
     /// to this read id, all of them are used.
@@ -255,7 +255,7 @@ pub struct InputBam {
     /// based on the other options that the user sets.
     #[clap(skip)]
     #[builder(setter(into, strip_option))]
-    pub region_bed3: Option<Bed3<i32, u64>>,
+    pub region_bed3: Option<Bed3<i32, u32>>,
     /// Only keep reads if they pass through the specified region in full.
     /// Related to the input `--region`; has no effect if that is not set.
     #[clap(long, default_value_t, requires = "region")]
@@ -468,7 +468,7 @@ impl TagState for RequiredTag {
 ///     .mod_prob_filter(ThresholdState::GtEq(0))
 ///     .trim_read_ends_mod(10)
 ///     .base_qual_filter_mod(10)
-///     .region_bed3(Bed3::<i32, u64>::new(4, 4000, 5000))
+///     .region_bed3(Bed3::<i32, u32>::new(4, 4000, 5000))
 ///     .build()?;
 ///
 /// # Ok::<(), Error>(())
@@ -485,7 +485,7 @@ impl TagState for RequiredTag {
 ///
 /// let input_options = InputModsBuilder::<RequiredTag>::default()
 ///     .mod_prob_filter(ThresholdState::GtEq(0))
-///     .region_bed3(Bed3::<i32, u64>::new(4, 4000, 5000))
+///     .region_bed3(Bed3::<i32, u32>::new(4, 4000, 5000))
 ///     .mod_region("chr3:4000-5000".into())
 ///     .build()?;
 ///
@@ -568,7 +568,7 @@ pub struct InputMods<S: TagState + Args + FromArgMatches + Default> {
     /// into a numeric contig id.
     #[clap(skip)]
     #[builder(setter(strip_option))]
-    pub region_bed3: Option<Bed3<i32, u64>>,
+    pub region_bed3: Option<Bed3<i32, u32>>,
 }
 
 impl<S: TagState + Args + FromArgMatches + Default> InputModsBuilder<S> {
@@ -631,11 +631,11 @@ pub trait InputModOptions {
 /// Retrieves options for region
 pub trait InputRegionOptions {
     /// returns region requested
-    fn region_filter(&self) -> &Option<Bed3<i32, u64>>;
+    fn region_filter(&self) -> &Option<Bed3<i32, u32>>;
     /// returns region requested but region in genomic string format
     fn region_filter_genomic_string(&self) -> Option<GenomicRegion>;
     /// sets region requested
-    fn set_region_filter(&mut self, _value: Option<Bed3<i32, u64>>);
+    fn set_region_filter(&mut self, _value: Option<Bed3<i32, u32>>);
     /// returns true if full overlap with region is requested as opposed to
     /// only partial overlap. defaults to false.
     fn is_full_overlap(&self) -> bool {
@@ -690,10 +690,10 @@ impl<S: TagState + Args + FromArgMatches + Default> InputRegionOptions for Input
     fn region_filter_genomic_string(&self) -> Option<GenomicRegion> {
         self.mod_region.clone()
     }
-    fn region_filter(&self) -> &Option<Bed3<i32, u64>> {
+    fn region_filter(&self) -> &Option<Bed3<i32, u32>> {
         &self.region_bed3
     }
-    fn set_region_filter(&mut self, value: Option<Bed3<i32, u64>>) {
+    fn set_region_filter(&mut self, value: Option<Bed3<i32, u32>>) {
         self.region_bed3 = value;
     }
 }
@@ -711,10 +711,10 @@ impl InputRegionOptions for InputBam {
     fn region_filter_genomic_string(&self) -> Option<GenomicRegion> {
         self.region.clone()
     }
-    fn region_filter(&self) -> &Option<Bed3<i32, u64>> {
+    fn region_filter(&self) -> &Option<Bed3<i32, u32>> {
         &self.region_bed3
     }
-    fn set_region_filter(&mut self, value: Option<Bed3<i32, u64>>) {
+    fn set_region_filter(&mut self, value: Option<Bed3<i32, u32>>) {
         self.region_bed3 = value;
     }
     fn is_full_overlap(&self) -> bool {
@@ -785,7 +785,7 @@ pub enum SeqDisplayOptions {
         /// Option to show modified bases as Z (or z depending on other options)
         show_mod_z: bool,
         /// Region over which sequence must be shown
-        region: Bed3<i32, u64>,
+        region: Bed3<i32, u32>,
     },
 }
 
@@ -907,7 +907,7 @@ mod input_mods_required_tag_tests {
             .mod_prob_filter(ThresholdState::GtEq(0))
             .trim_read_ends_mod(10)
             .base_qual_filter_mod(10)
-            .region_bed3(Bed3::<i32, u64>::new(4, 4000, 5000))
+            .region_bed3(Bed3::<i32, u32>::new(4, 4000, 5000))
             .build()
             .unwrap();
     }
@@ -951,8 +951,8 @@ mod input_bam_tests {
     fn input_bam_builder() {
         let _: InputBam = InputBamBuilder::default()
             .bam_path(PathOrURLOrStdin::Path("/some/path/to/bam.bam".into()))
-            .min_seq_len(30000u64)
-            .min_align_len(20000i64)
+            .min_seq_len(30000u32)
+            .min_align_len(20000u32)
             .read_id("some-id")
             .read_filter("primary_forward,secondary_forward".into())
             .sample_fraction(F32Bw0and1::one())
@@ -1074,7 +1074,7 @@ mod input_bam_tests {
         // When region_bed3 is already set and region is None, should not convert
         let mut input_bam = InputBam {
             region: None,
-            region_bed3: Some(Bed3::<i32, u64>::new(1, 1000, 2000)),
+            region_bed3: Some(Bed3::<i32, u32>::new(1, 1000, 2000)),
             ..Default::default()
         };
 
@@ -1103,7 +1103,7 @@ mod input_bam_tests {
         // When both region and region_bed3 are set, should error
         let mut input_bam = InputBam {
             region: Some(GenomicRegion::from_str("chr2:1500-2500").unwrap()),
-            region_bed3: Some(Bed3::<i32, u64>::new(1, 1000, 2000)),
+            region_bed3: Some(Bed3::<i32, u32>::new(1, 1000, 2000)),
             ..Default::default()
         };
 
@@ -1126,12 +1126,12 @@ mod input_region_options_tests {
 
     #[derive(Debug, Default)]
     struct TestInputRegionOptions {
-        region_filter: Option<Bed3<i32, u64>>,
+        region_filter: Option<Bed3<i32, u32>>,
         region_filter_genomic_string: Option<GenomicRegion>,
     }
 
     impl InputRegionOptions for TestInputRegionOptions {
-        fn region_filter(&self) -> &Option<Bed3<i32, u64>> {
+        fn region_filter(&self) -> &Option<Bed3<i32, u32>> {
             &self.region_filter
         }
 
@@ -1139,7 +1139,7 @@ mod input_region_options_tests {
             self.region_filter_genomic_string.clone()
         }
 
-        fn set_region_filter(&mut self, value: Option<Bed3<i32, u64>>) {
+        fn set_region_filter(&mut self, value: Option<Bed3<i32, u32>>) {
             self.region_filter = value;
         }
     }
@@ -1173,7 +1173,7 @@ mod validate_builder_functions {
         let _: InputBam = InputBamBuilder::default()
             .bam_path(PathOrURLOrStdin::Path("/some/path.bam".into()))
             .region("chr1:1000-2000".into())
-            .region_bed3(Bed3::<i32, u64>::new(0, 1000, 2000))
+            .region_bed3(Bed3::<i32, u32>::new(0, 1000, 2000))
             .build()
             .unwrap();
     }
@@ -1237,7 +1237,7 @@ mod validate_builder_functions {
         let _: InputMods<OptionalTag> = InputModsBuilder::<OptionalTag>::default()
             .mod_prob_filter(ThresholdState::GtEq(0))
             .mod_region("chr1:1000-2000".into())
-            .region_bed3(Bed3::<i32, u64>::new(0, 1000, 2000))
+            .region_bed3(Bed3::<i32, u32>::new(0, 1000, 2000))
             .build()
             .unwrap();
     }
@@ -1249,7 +1249,7 @@ mod validate_builder_functions {
             .tag(RequiredTag::from_str("m").unwrap())
             .mod_prob_filter(ThresholdState::GtEq(0))
             .mod_region("chr1:1000-2000".into())
-            .region_bed3(Bed3::<i32, u64>::new(0, 1000, 2000))
+            .region_bed3(Bed3::<i32, u32>::new(0, 1000, 2000))
             .build()
             .unwrap();
     }
